Trying to replace multiplatform ReKotlin with the standard one.
In order to revert this change, I'll have to revert to the present
commit, then change the dependencies back manually, and then fix it from
there.

This is not going to fix anything.  I'm trying to fix an issue with
Rekotlin Router, and I'm just now reminded that these two components
are separate from one another.